requirements.txt
langchain
langchain-openai
langchain-community
tavily-python
python-dotenv
pydantic